+++ vgrind/regexp.c
@@ -30,7 +30,7 @@ static void expconv(void);
 
 #define isidchr(c)	\
 		(isalnum(c) || ((c) != NIL && strchr(l_idchars, (c)) != NIL))
-#define makelower(c)	(isupper((c)) ? tolower((c)) : (c))
+#define makelower(c)	(isupper((int)(c)) ? tolower((int)(c)) : (c))
 
 /*  STRNCMP -	like strncmp except that we convert the
  *	 	first string to lower case before comparing
@@ -434,7 +434,7 @@ expmatch (
 			/* not optional and we still matched */
 			return (NIL);
 		    }
-		    if (!isidchr(*s1))
+		    if (!isidchr((int)*s1))
 			return (NIL);
 		    if (*s1 == '\\')
 			escaped = escaped ? FALSE : TRUE;
